refactor: change in directory hierarchy
