+++ packages/hoppscotch-js-sandbox/src/__tests__/testing/envs/get.spec.ts
@@ -1,8 +1,9 @@
+import "@relmify/jest-fp-ts"
 import * as TE from "fp-ts/TaskEither"
 import { pipe } from "fp-ts/function"
-import { execTestScript, TestResponse, TestResult } from "../../../test-runner"
 
-import "@relmify/jest-fp-ts"
+import { execTestScript } from "../../../test-runner/node-vm"
+import { TestResponse, TestResult } from "../../../types"
 
 const fakeResponse: TestResponse = {
   status: 200,
