+++ packages/hoppscotch-js-sandbox/src/__tests__/testing/envs/resolve.spec.ts
@@ -1,6 +1,8 @@
-import { pipe } from "fp-ts/function"
 import * as TE from "fp-ts/TaskEither"
-import { execTestScript, TestResponse, TestResult } from "../../../test-runner"
+import { pipe } from "fp-ts/function"
+
+import { execTestScript } from "../../../test-runner/node-vm"
+import { TestResponse, TestResult } from "../../../types"
 
 const fakeResponse: TestResponse = {
   status: 200,
