+++ packages/hoppscotch-js-sandbox/src/__tests__/testing/expect/toBe.spec.ts
@@ -1,7 +1,9 @@
+import "@relmify/jest-fp-ts"
 import * as TE from "fp-ts/TaskEither"
 import { pipe } from "fp-ts/function"
-import { execTestScript, TestResponse } from "../../../test-runner"
-import "@relmify/jest-fp-ts"
+
+import { execTestScript } from "../../../test-runner/node-vm"
+import { TestResponse } from "../../../types"
 
 const fakeResponse: TestResponse = {
   status: 200,
